refactory: member 도메인 이동
