+++ backend/src/main/java/toonpick/app/auth/handler/LoginSuccessHandler.java
@@ -12,7 +12,7 @@ import org.springframework.security.web.authentication.SimpleUrlAuthenticationSu
 import org.springframework.stereotype.Component;
 import toonpick.app.auth.jwt.JwtTokenProvider;
 import toonpick.app.auth.token.TokenService;
-import toonpick.app.service.MemberService;
+import toonpick.app.member.service.MemberService;
 
 import java.util.stream.Collectors;
 import java.io.IOException;
