+++ backend/src/main/java/toonpick/app/auth/user/CustomOAuth2UserDetails.java
@@ -2,7 +2,7 @@ package toonpick.app.auth.user;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
-import toonpick.app.dto.MemberDTO;
+import toonpick.app.member.dto.MemberDTO;
 
 import java.util.ArrayList;
 import java.util.Collection;
