+++ backend/src/main/java/toonpick/app/dto/WebtoonReviewDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import toonpick.app.member.dto.MemberDTO;
 
 import java.time.LocalDateTime;
