+++ backend/src/main/java/toonpick/app/entity/ReviewLike.java
@@ -13,6 +13,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import toonpick.app.member.entity.Member;
 
 @Entity
 @Getter
