+++ backend/src/main/java/toonpick/app/entity/WebtoonReview.java
@@ -5,6 +5,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import toonpick.app.member.entity.Member;
 import toonpick.app.webtoon.entity.Webtoon;
