+++ backend/src/main/java/toonpick/app/member/service/MemberFavoriteWebtoonService.java
@@ -1,15 +1,15 @@
-package toonpick.app.service;
+package toonpick.app.member.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import toonpick.app.webtoon.dto.WebtoonDTO;
-import toonpick.app.entity.Member;
-import toonpick.app.entity.MemberFavoriteWebtoon;
+import toonpick.app.member.entity.Member;
+import toonpick.app.member.entity.MemberFavoriteWebtoon;
 import toonpick.app.webtoon.entity.Webtoon;
 import toonpick.app.exception.ResourceNotFoundException;
 import toonpick.app.webtoon.mapper.WebtoonMapper;
-import toonpick.app.repository.MemberFavoriteWebtoonRepository;
-import toonpick.app.repository.MemberRepository;
+import toonpick.app.member.repository.MemberFavoriteWebtoonRepository;
+import toonpick.app.member.repository.MemberRepository;
 import toonpick.app.webtoon.repository.WebtoonRepository;
 
 import java.time.LocalDateTime;
