+++ backend/src/test/java/toonpick/app/unit/entity/MemberTest.java
@@ -2,7 +2,7 @@ package toonpick.app.unit.entity;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import toonpick.app.entity.Member;
+import toonpick.app.member.entity.Member;
 
 import static org.assertj.core.api.Assertions.assertThat;
