+++ backend/src/test/java/toonpick/app/unit/service/WebtoonReviewServiceTest.java
@@ -7,13 +7,13 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 import toonpick.app.dto.WebtoonReviewCreateDTO;
 import toonpick.app.dto.WebtoonReviewDTO;
-import toonpick.app.entity.Member;
+import toonpick.app.member.entity.Member;
 import toonpick.app.webtoon.entity.Webtoon;
 import toonpick.app.entity.WebtoonReview;
 import toonpick.app.entity.enums.Platform;
 import toonpick.app.entity.enums.SerializationStatus;
 import toonpick.app.exception.ResourceNotFoundException;
-import toonpick.app.repository.MemberRepository;
+import toonpick.app.member.repository.MemberRepository;
 import toonpick.app.webtoon.repository.WebtoonRepository;
 import toonpick.app.repository.WebtoonReviewRepository;
 import toonpick.app.service.WebtoonReviewService;
